Fix : Entity ID int형에서 Long형으로 변경
JPA에서 ID 엔티티는 int 대신 Long형을 지원한다고하여 데이터베이스의 ID 엔티티 자료형을 Long으로 변경함.

+++ src/main/java/sch/cqre/api/model/entity/CommentEntity.java
@@ -20,13 +20,13 @@ public class CommentEntity {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Id
 	@Column(name = "comment_id")
-	private int commentId;
+	private Long commentId;
 	@Basic
 	@Column(name = "post_id")
-	private Integer postId;
+	private Long postId;
 	@Basic
 	@Column(name = "author_id")
-	private Integer authorId;
+	private Long authorId;
 	@Basic
 	@Column(name = "comment_content")
 	private String commentContent;
