+++ src/main/java/sch/cqre/api/model/entity/PostEntity.java
@@ -20,10 +20,10 @@ public class PostEntity {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Id
 	@Column(name = "post_id")
-	private int postId;
+	private Long postId;
 	@Basic
 	@Column(name = "author_id")
-	private int authorId;
+	private Long authorId;
 	@Basic
 	@Column(name = "post_title")
 	private String postTitle;
