+++ src/main/java/sch/cqre/api/model/entity/SupplyEntity.java
@@ -19,7 +19,7 @@ public class SupplyEntity {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Id
 	@Column(name = "supply_id")
-	private int supplyId;
+	private Long supplyId;
 	@Basic
 	@Column(name = "name")
 	private String name;
